Makes the counting sampler accurate to 100 decisions and more fair
Previously, the counting sampler biased the first trace ids. For
example, if you had a sample rate of 50%, 5K traces would be sampled,
then 5K would not be. This would result in a poor experience for low
traffic sites. This randomizes the decisions, so that they are more
fair, and restores the granularity to 100 decisions.