Fixes crash where there's a race between storage and query components
Formerly, wiring `ZipkinQueryApiV1` would crash the app if a storage
connection was down. In other words, it was doing I/O in the
constructor. The storage component is lazy, so the way to fix this is to
defer access until it is needed.

This also adds a test so that it is found before docker packaging.