Deploys snapshots and versions from master
Since our release task makes tags, not branches, version commits end up
on master. These deploy, so we can't also build the tag, too.

This consolidates code around deployment to the master branch, and uses
the maven version to tell if it should be synchronized to maven central
or not.