Drops all dependencies on the scala project
This drops all build dependencies on the scala project, reducing build
time and flakes due to Twitter repository returning 503 errors.

The most notable change is this drops the "interop" module, which was
used to develop the project. Since zipkin-java is full featured now, and
there is little change in zipkin-scala anymore, this is no longer
pulling its weight.