Extracts auto-configuration modules from zipkin-server
This decouples zipkin-server from ui, storage and collector modules.
While there remains an optional dependency, this is only a mechanism to
include these modules in the fat jar.

To do this, it uses the AutoConfigure technique from Spring Boot,
notably registering a manifest for each component to export.

To completely decouple, some work was needed. Particularly health check
registration needed to become more generic, and occur at the very end of
the application bootstrap so as to not interfere with initialization.

Finally, this clears up a long-standing confusion where we named a
storage component jdbc, eventhough we only test it with mysql and all
the flags are named mysql. This settles on mysql as for example
PostgreSQL would likely need different syntax.