Removes tracegen
Tracegen was a tool we used to use to populate zipkin when someone had
no instrumented apps. It would create a number of traces based on latin
words.

Since then, we have increasingly sophisticated and relevant self-tracing
setup, which in the java project extends all the way to cassandra.

Moreover, we have spigo, which now includes zipkin scripts to create
large amounts of realistic traces, and instrumentation-specific examples
from Brave and Sleuth.

Considering the better alternatives, we should stop maintaining tracegen

https://github.com/adrianco/spigo
https://github.com/openzipkin/brave-resteasy-example
https://github.com/spring-cloud-samples/brewery