Extracts DeduplicatingExecutor, preventing redundant Cassandra indexing (#1157)
The previous code had a mechanism to reduce writes to two indexes:
`service_name_index` and `service_span_name_index`. This mechanism would
prevent writing the same names multiple times. However, it is only
effective on a per-thread basis (as names were stored in thread locals).

In practice, this code is invoked at collection, and collectors have
many request threads per transport. By changing to a shared loading
cache, we can extend the deduplication to all threads. By extracting a
class to do this, we can test the edge cases and make it available for
future work, such as the other indexes.