Optimizes out cassandra index inserts that don't improve query results
For example, a trace that fans out into N spans against the same service
will end up with a fixed amount of cassandra inserts as opposed to a
function of N.

This removes extra requests when a collector receives a large amount of
span data for a single trace. That could be the case when one or more of
the following are true:

* instrumentation bundles all local spans in a trace into one message
* traces are routed consistently to a single storage component
* there's only one collector

The overhead of this feature with defaults should be tens of megabytes
heap. Disable by setting `CassandraStorage.Builder.indexCacheMax(0)`