Conditionally supports pre-aggregated dependencies in MySQL
This adds the table `zipkin_dependencies`, which is read when present
to return pre-aggregated dependencies (from the zipkin-dependencies job)

If there is no data in the `zipkin_dependencies` table, results will be
aggregated on-demand. This allows backwards compatibility for those who
haven't added the new table, or run the spark job yet.