Decouples maximum lookback from default lookback (#1212)
This does a number of things to improve performance.

QUERY_LOOKBACK prevents the server from having to construct queries that
are longer than indexes. This value was used by default which led to a
couple major problems.
* When looking for performance problems, you aren't usually looking for
7 day old data, and even if you were, it wouldn't fit into a limit of
10. Changing to a much smaller amount was requested in #1210
* Cassandra duration queries convert into lookback/1hr requests. A
default of 7 days means 168 requests. This led to most users being
unable to use the duration query function.
* By changing to 1hr, our 10 trace limit is very likely to apply, and in
the case of Cassandra, we will only pull a single bucket.

In #1185, we know we want to change to daily granularity. By switching
lookback to 1 day (decoupled from the 1hr lookback on traces), we better
prepare for that change.

I also noticed that we document queryLimit to 10, but config.js says 15.
Regardless of the correct value, they should be aligned.