Moves zipkin-core language level to JRE 6
This is an api compatible change to allow code sharing between zipkin
server and instrumentation, and without having to publish 2 jar files.

We accidentally broke the ability to make agent-based instrumentation by
compiling our model classes against Java 7 bytecode. Particularly, this
causes a regression in Brave.

To move back to Java 6 bytecode, we had to do a few things:
* replace moshi internal dependency with gson, which is java 6 compliant
* remove okio dependency and copy its Base64 class
  * incidentally a class it also copied!
* do variable-size buffer writes w/ OutputBuffer instead of okio
* fix a few nags about diamond ops etc.