elasticsearch: add aws request signing capability
When connecting to an aws elasticsearch managed cluster, requests for
any non-open domain must be signed with the requester's IAM
credentials. This change looks for AWS-managed elasticsearch in the
connection URL, and, if present, adds a request signing interceptor
to allow access.

Additionally, it is sometimes convenient to refer to the domain by
human-readable name rather than machine-unique "endpoint": an
additional set of configuration parameters have been added to support
this case.