Cleans up InternalElasticsearchClient and fixes aggregation bug (#1313)
This cleans up InternalElasticsearchClient in preparation of
Elasticsearch 5 work.

Formerly, the client implementation was required to search the response
of a request for specific paths of aggregation bucket keys. This
required an intermediary to allow the same response to be read for two
paths. Since implementations are trusted, the change is that we search
for "any" bucket keys. This is simpler to test and requires less
coordination.

On that note, there was a bug in the aggregation logic, where annotation
aggregations weren't treated as nested. This fixes the bug and rewrites
the test to expose it in the future.

In order to plug-in a different span indexer, I needed to re-organize
that code. Now, there's a BulkSpanIndexer type which can be exteded as
needed.

See #1312