Compresses indexes to search based on date expression
This adds `IndexNameFormatter.indexNamePatternsForRange` which
compresses api requests by using wildcard expressions for months or
years.

For example, if begin is 2016-11-30 and end is 2017-01-02, the result
will be 2016-11-30, 2016-12-*, 2017-01-01 and 2017-01-02 as opposed to
a set of >30 individual index names.

As a side effect, we no longer need special http configuration as we
won't create request lines larger than 4096 on account of index names.