be more forgiving of IPv4-mapped addresses in json
Currently Zipkin checks that all IPv6 addresses are 16 bytes long.
However, if an IPv4-mapped addresses such as `::ffff:192.168.1.128` is
given then InetAddress.getByName will return an Inet4Address and the
assertion will fail.

Instead accept the address, but store it as ipv4 to maintain all
existing assertions.  This absolves clients of reconciling the nuances
of IPv4-mapped addresses in their environments (for example, in node
`net.isIPv6('::ffff:192.168.1.128')` is true) with Java's.