Moves ES 5 docker to new base image; deletes broken transport 5 test
This moves the ES 5 docker tests to our new base image, which
significantly reduces first-time download size (particularly helpful for
those on slow networks).

It also deletes the `ElasticsearchV5TransportTest` as it will never work
since the 2.x protocol is incompatible with 5.x and we can only use one.

Finally, it reformats imports so they match the others.