Add logs button on the trace page.
When troubleshooting our system, we often need to check
if there were logs. Since a good pratice is to add the
trace id to your logs, it seems really natural to have
a button to open a link to query your logs with this id.
Also, people generally have kibana to search for specific
logs (or at least have a query-able log service).

The button is deactivated by default (if you don't have a log
query service or the trace id in the logs) and it's simply
activated by adding LOGS_URL="http://company.com/logs?query={traceId}".
The traceId will be contextually replaced.