Adds ES_MAX_REQUESTS to control maximum in-flight Elasticsearch requests
The http client used for Elasticsearch is exclusively for Elasticsearch.
However, it was using defaults for web sites, which constrained it to 5
in-flight requests per host. This raises the default to the same as max
requests, which is 64. It also opens `ES_MAX_REQUESTS` to adjust that
count.

The intended behavior is that there will be less contension in a server
that is both storing spans and serving api requests. It should also
allow higher throughput to the Elasticsearch cluster.

Fixes #1449