Decouples zipkin-elasticsearch-http from Elasticsearch 2.x libraries
Formerly, the http Elasticsearch module had a hard dependency on
Elasticsearch 2.x libraries. This was simply work left unfinished, as
the effort to completely decouple was high.

This change completes the separation of the http elasticsearch code
from the prior codebase. You can now drop in `ElasticsearchHttpStorage`
where you previously used `ElasticsearchStorage`. You can now use
Elasticsearch without a guava or native library dependency.

Server dependencies and default configuration are left unaffected except
that there is no more sharing between the auto-configuration classes of
the native transport vs the http transport. This means that if you make
a custom server with `zipkin-autoconfigure-storage-elasticsearch-http`,
you won't transitively pull in Guava or Elasticsearch 2.x libraries.

The old `HttpClientBuilder` is now deprecated and as mentioned above, no
longer used. Until it is removed, there is an optional dependency on the
native code in `zipkin-storage-elasticsearch-http`. Beyond clutter, and
increasing test duration, there's little impact to this.