Do not add incorrect dependencies when processing partial/broken traces
The current dependency building logic uses the first spen as root for span
trees that are missing a span with null parentId.

So for example given this span tree (where all spans are server spans):

    1: {traceId: 1, spanId: 1, serviceName: A}
    2: {traceId: 1, spanId: 2, serviceName: B, parentId: 1}
    3: {traceId: 1, spanId: 3, serviceName: C, parentId: 1}

if the first span is not received, the dependency graph will contain a B -> C
dependency, which is obviously wrong.

This commit uses an empty node as the tree root for this case instead of using
the first server span as the parent.