Add an extra field to mark the synthetic root node
It's hard to grep for null checks; having a special sentinel value requires
either to pass it to constructTree() (for the type-safe version) or using a
statically-allocated Object instance (for the exploiting-erasure version).

The first version is clunky, the second version will make code that calls

    Span value = node.value();

fail with a Classcast exception unless it first calls the
isSyntheticRootForPartialTree() accessor.

I don't particularly like the current implementation, I just like it a bit more
than the two options above.