Drops Elasticsearch Native transport (#1547)
This drops the Elasticsearch Native transport, which allows us to focus
efforts on the http variant. It also stops us from interfering with
Spring Boot autoconfiguration of Elasticsearch. Finally, it speeds up
the build as related docker tests are no longer needed.

To help in transition, this implicitly converts server configuration
pointing to the native port 9300 towards the http port 9200 after
logging a warning. This is an attempt to not break existing users.