Decouples v2 types from v1 types, in preparation of a new package
This moves all the code needed for v2 types into zipkin.internal.v2
without any references to code outside the package. To reduce
duplication, "v1" code can reference v2 types (as they need to in order
to covert for example). Once this is in, we can consider refactoring out
a v2 module.