Teaches the v2 tree how to encode v1 format
The zipkin v2 library is intended to be used by itself in tracers. Since
tracers cannot control the span transport, it likely will need to write
the old (historical) json format for at least months. To avoid requiring
dual versions of jars to accomplish this, we add an encoder that writes
json v1 format given a v2 span.

Note: we don't decode as the library problem is easier to manage on the
server side. For example, zipkin-server will need both versions of the
jars for things like collection for a while.