Require JDK 8 with enforcer plugin
To avoid confusing failure messages when a version of JDK that doesn't work with the code is used, this requires JDK 8 using the enforcer-plugin, which will output a sensible error message in case an incompatible JDK version is used. Notes the JDK 8 requirement in the README where instructions for building from source are.