zipkin-ui: include minimal header on page load
One of the reasons why the zipkin-ui _feels_ slow is because when you click a link, the page is first rendered as blank, and then re-rendered with the page content.

This can be improved significantly by rendering the header as part of the original `index.html` page load. This is much less visually disruptive, and makes everything feel a lot snappier.

I copied the navbar code from `layout.mustache`.