zipkin-ui: relative interval-based lookback (#1872)
the default start-ts and end-ts elements in the form are imo
a little difficult to use. commonly i am usually interested
in some sort of relative interval, or lookback.

this patch replaces the start-ts and end-ts fields with a
pre-configured dropdown of "1 hour", "3 hours", "6 hours",
"12 hours", "1 day", "2 days", and "7 days".

there is also a "custom" option which reveals the
start-ts and end-ts fields.