Defined http.route key and ensures server span name is preferred (#1924)
Especially when using http route based span names, we have to be careful
to not pick a bad name. In shared spans, for example, the UI prefers
server-side's name of the service. However, we have no name priority
logic, so it is arbitrary. This ensures the server's span name wins in a
race.

See https://github.com/openzipkin/zipkin/issues/1874