Deprecates EnableZipkinServer to explain custom servers are unsupported
Especially lately, we have had a large number of people having problems
with unnecessarily custom servers. Some are due to not knowing Sleuth's
stream server is obviated by our rabbit support. Some are due to blogs
which unfortunately recommend starting Zipkin in the IDE via a custom
server.

Through discussion, we decided the easiest way to let users know custom
servers are unsupported is by deprecation. Deprecation shows up in the
IDE and will alert those doing blogs or otherwise that they are
suggesting discouraged practice. It also sends a clear signal to those
who need to make custom servers that while doing so is possible, it is
something the customizer needs to accept support reponsibility of.