Changes index page to order services descending by max span duration (#2234)
I figured out what "max" was in the index page. It corresponds to the
longest duration span in the trace, as attributed to a specific service.

This orders the search results descending by the most likely culprit,
using relevant time units to prevent minutes from showing as huge counts
of milliseconds, or short traces showing up as zero.