Conditionally filters when strictTraceId == true
In order to read data written before switching from strictTraceId == false, we
filter client side even when strictTraceId == true. This broke some behavior
Netflix relied on that allowed multi-tag based query on "fit" experiments.

There are a few ways to address this, including adding another flag to say if
you have any old data or not. In this case, you'd never want to filter client-
side. OTOH introducing a new flag is burdensome from a documentation POV.

This addresses the concern in a different way. The only reason we would have to
filter client side, is if there is a clash on the lower 64-bits of the trace ID.
Knowing this, we can achieve the same with no configuration just by looking for
a clash and doing no filtering if there is no clash.