Unsupports v1 trace upload
Javascript conversion of v1 trace data to v2 format is a fairly large
task. We will likely do such a thing in the zipkin-js library vs here as
the code is substantial and not specific to the UI.

This unsupports the trace viewer screen for v1 format. A workaround is
to use a local zipkin server to convert the data meanwhile.

Fixes #2269