Fixes display order of spans in a trace (#2313)
@drolando noticed a bug in our presentation of spans when there are
parallel requests. Basically, the children were not placed under their
parents, which prevented the collapse feature from working. We didn't
notice because we only had test data for serial trace and we also
created bad data in a test (we tested for the wrong traversal order).

This switches to depth first traversal, which fixes the problem.