Do not try to create RabbitMQ queue if one already exists
RabbitMQ may be managed by a team that will create queues with additional configuration beyond the default values the RabbitMQCollector would use. Before the collector would fail to start because it unconditionally tried declaring a queue but one with the same name and different configuration already existed.

This also cleans up the channel used to check and maybe create the queue, which used to stick around even though it would never be used again.

Fixes #2068