Works around encoding bug in zipkin-ruby by leniently parsing
In zipkin-ruby (zipkin-tracer up to 0.32.3), the "sa" annotation was
incorrectly encoded as "1" not `true`. This likely was due to the fact
that thrift encodes boolean as 1. This caused the remote endpoint to be
mistaken for a local endpoint as the parser thought "sa" was a normal
string tag with the local endpoint attached. The impact was
mis-rendering and also service graph aggregation would likely be
incorrect in some cases as well.

Definitely, zipkin-ruby should update as even if we leniently parse,
clones will not likely be as tolerant and this could hurt things.
However, it is best to be lenient as it is cheap and low impact. There
is unlikely any valid tag named "sa" "ca" or "sa" used for a purpose
besides the remote endpoint.

Fixes #2414