WIP: Offer go back to classic when switched to lens
In a previous change, we offered to switch to Lens. This offers to
switch back, based on the same cookie (`lens`). This should help
reduce risk of people not wanting to deploy.

Note: this needs css work and also I'm not sure why the page isn't
reloading completely. I would like some help finishing this.