Fixes some lens UI bugs
Service names are now alphabetically sorted. The API seems to return
them out of order, so let's sort them just to be sure.

Added missing space between span count and the "span" word.

Timestamps in the span details view now have microsecond precision.
moment.js doesn't support that, so I had to append the last 3 digits of
the timestamp to the human readable string.

CSS is weird. Apparently it's enough to set `white-space: normal` to
make long text wrap correctly in the span details view. I also tested
with a single very long word and if that doesn't fit in the screen it
gets split automatically.