Fixes logging and metrics for collectors
Before, we were inconsistent about incrementing metrics, and in worst
case, we double-incremented. This adds tests for the messaging based
collectors and backfills tests to make the base Collector class 100%
coverage.

Separately, we have recently changed our web endpoint to produce its
own error messages to the client. In other words, we no longer rely on
(often unnecessary) wrapping of exceptions to correct the message being
sent back. This changes the logic around message formatting so it is
only used when logging is enabled.