Verifies ES 5.x with unit tests as opposed to integration tests (#2560)
Our builds currently take too long due in part to checking 3 different
versions of Elasticsearch. Elastic no longer supports version 5, so our
support of it is "best efforts". Accordingly, integration testing v5.x
knowing we break the build on timeouts is not the best call.

This increases the unit test coverage of version specific code to 100%
and removes the Elasticsearch 5.x integration tests.