Unhooks protocols that are usually synchronous from storage commands
http and gRPC protocols usually hold clients. We want to hold them until
messages are parsed, but not later. This employs an executor to ensure
any blocking commands needed to get the storage command enqueued do not
block the request.

This also fixes some lack of logging where nio is used.

Fixes #2715