Instruments when STORAGE_THROTTLE_ENABLED=true
Before, we couldn't see the effects of self-tracing when throttling
was enabled. This change simplifies some of the code in the throttling
feature such that it uses a simpler to instrument `Executor` to invoke
commands. This also makes explicit some tracing we are trying to avoid,
notably anything spawned by the tracer itself.