Makes throttling parameters obvious by controlling exceptions
Before, I was a bit confused about what the different rejected errors
implied, both looking at the code and tests. This takes a more literal
approach, considering the invocation is only from the server. Instead,
this says which variables have been crossed in the exception message.

Secondarily, this uses a single exception for each of the cases in order
to reduce load (of creating exceptions) when a server is constantly
overloaded.