Falls back to simple name and stops prefixing rejected ES exceptions
Armeria has a number of constant throwables, which neither have a stack
trace, nor a message. This switches to use these while we figure out if
there is a better way to unveil an underlying http connection problem
the health-check sees, but a health-checked client would not.

The approach is used all over in zipkin, most notably Brave.

This also stops prefixing RejectionException messages also with the word
as it is annoying to read it twice in toString.

Most importantly, this unwraps RejectionException in health check as it
doesn't effectively describe the cause of the problem, rather obfuscates
it.

Before:
```bash
curl -s localhost:9411/health
{"status":"DOWN","zipkin":{"status":"DOWN","details":{"ElasticsearchStorage{initialEndpoints=https://localhost:9200, index=zipkin}":{"status":"DOWN","details":{"error":"java.util.concurrent.RejectedExecutionException: Rejected execution: null"}}}}}
```

Now:
```bash
curl -s localhost:9411/health
{"status":"DOWN","zipkin":{"status":"DOWN","details":{"ElasticsearchStorage{initialEndpoints=https://localhost:9200, index=zipkin}":{"status":"DOWN","details":{"error":"com.linecorp.armeria.common.ClosedSessionException"}}}}}
```