Makes Lens default, removing the dependency on the classic UI (#2971)
For well over a year, we've been working on Lens, most notably @tacigar.

In fact a year ago, we added a "Try Lens" button and addressed a lot of
feedback from that, again largely thanks to @tacigar.

It isn't fair to hold the project hostage until some notion of
perfection happens. It is also unfair for people to pick certain issues
to block on, yet not volunteer to help. It is of my opinion we stop
holding ourselves hostage to a maintenance pit and remove the burden and
confusion of two UIs.

This concedes the following just won't work until someone volunteers to
fix them. Waiting a year hasn't done it, so waiting more won't either.

* ZIPKIN_UI_BASEPATH for reverse proxy https://github.com/openzipkin/zipkin/issues/2519
* ZIPKIN_UI_LOGS_URL https://github.com/openzipkin/zipkin/issues/2491
* /config.json config overrides https://github.com/openzipkin/zipkin/issues/2969
* Chinese language support https://github.com/openzipkin/zipkin/issues/2970