Dynamic elasticsearch security file
Load elasticsearch credentials from a properties file in a fixed
rate interval. There are two parameters to control it:
  * security-file-path: An absolute path of file.
  * security-file-refresh-interval-in-second: Refresh interval in
    seconds.

Notic: Before the crendentails stored in properties file are loaded
into heap, username/password parameters are as the default value.
If one of them is a empty string, basic authentication is disable.
That means "Authorization" header is absent in http request.

Signed-off-by: Gao Hongtao <hanahmily@gmail.com>