Refactors DelayLimiter in preparation of re-use in cassandra v1 (#3136)
This refactors code relating to DelayLimiter so that it is easier to
reuse later. What will happen in a separate PR (or what will be
attempted) is change the limiter so that it can run a predicate instead
of an exact match. For example Cassandra v1's Indexer currently mutes
index insertions which are logically equivalent to data already written.

Concretely, a service name that appears 3 times in the same trace,
results in two writes (earliest and latest). If it appears again, and
between those timestamps, that write is muted as it is redundant.