Stops using prepared or query builder statements for simple queries (#3156)
Before, we needlessly prepared service name query, which would have no
parameters. While, this is being followed up below, it is unlikely that
preparing is better than directly executing the statement:
https://groups.google.com/a/lists.datastax.com/forum/#!topic/java-driver-user/d6wLkH3xDLI

More importantly, there is a bug in the cassandra driver >v3.6 which
crashes on `QueryBuilder` statements that have no parameters. This
uses a simple statement for that instead (healthcheck). Moreover, the
log category that shows connection failures is now suggested in our
docs.

See https://github.com/datastax/java-driver/pull/1138 for the NPE