Refactors Cassandra queries so they are cheaper and easier to migrate
Before, we used labels a lot, and generally a lot of query builder code.
This switches most to positional syntax, which is cheaper and easier to
migrate to cassandra driver 4.

This also reformats everything to make the migration feel like less
change.

Finally, this consolidates as much code as possible, notably repetitive
inserts and queries, again to reduce the heft needed to upgrade later.