Removes Cassandra querybuilder dependency
This removes the querybuilder dependency in efforts to reduce size.

While this saves a couple hundred K, it doesn't change things further as
datastax core lib depends still on a 2.8M shaded guava jar.

I've raised a help request on that for now, but reducing the API surface
area to only core libs should help anyway.

https://groups.google.com/u/1/a/lists.datastax.com/g/java-driver-user/c/aN7NkxX_yQk