Deprecates Cassandra v1 schema for removal in Zipkin 2.23
The "cassandra3" storage type relies on Cassandra v3.11.3+. This has
been out for over 2 years. We will soon need to support v4 which is
coming out of beta soon. Maintenance on the v1 schema has been immense
and I don't think volunteers would do any more refactors of it. It is
time to deprecate and remove the v1 schema.

"cassandra3" uses UDTs and optimized SASI indexing.  While SASI is
marked experimental, it is enabled by default in recent Cassandra
versions. Some clones of Cassandra may choose to not support SASI, but
this does not mean we must maintain an ancient schema. Those who must
use a clone can disable search, which avoids SASI.

It is likely that when Cassandra 4 comes out, we will use SAI indexing
instead with some migration instructions. However, SASI as used today
has been used in production for 2 years. Even if the feature is marked
experimental, it is well practiced here.