Changes base images to pull GitHub Container Registry instead of Docker Hub
This is the first step in removing a build outage risk due to Docker Hub
pull limits. This changes the base images used to something not Docker Hub,
mostly GitHub Container Registry except an edge case for our base NGINX
layer.

Notably, this does not fix zipkin-builder as that will be done in a
separate PR.

See https://github.com/openzipkin/docker-java/pull/33