Stops using zipkin-builder Docker image
This stops building a zipkin-builder image. This means that you can't
build Zipkin from source using only Docker.

Ex. a variant of the following won't work for source, only releases from
Maven central:
```bash
$ docker build https://github.com/openzipkin/zipkin
```

This implies Docker Hub auto-builds won't work for source releases,
which is not an issue as we don't use it anymore, and also Docker Hub
quota is problematic.

The pressure to remove source builds became strong when we started to
support multi-architecture builds. It is extremely time consuming to
attempt to run Maven inside of Docker, with an arm64 emulator. This
displaces effort that is sorely needed elsewhere and generally makes
the build unapproachable.

Since the Zipkin server itself is platform independent Java and client-
ide Javascript, there's no architectural impact to re-using the same
binaries for N arch builds. For example, this reworks the GitHub Action
validator to only build Zipkin once and use it for all the containers
to check.

Fixes #3271