Updates to latest dependencies and deletes invoker tests (#3285)
We originally supported server extensions, and so had Maven invoker
tests which isolate the classpath to hit a few areas. This deletes those
tests as we now use Docker for integrated server tests. This speeds up
the non-docker build quite a bit.

This also prunes the dependencies used in zipkin-server as the default
spring boot starter for test pulls in different versions of common
libraries such as assertj and jupiter. The only cost is maintaining the
spring-test version, which is only used in a few test runners, so not
particularly fragile.