Consolidates GH action build to a single file
This consolidates the two build processes into one flow, containing all
previous functionality, but less build time.

This builds the server on ubuntu, os/x and windows. If ubuntu, the flow
will continue into building and testing all our Docker images.

Something new is that this skips out if the change is documentation-only.